folderMQ - changed written file names to include "in_progress" until write operation it is complete

+++ engine/pubSub/core/folderMQ.js
@@ -1,11 +1,10 @@
 
 var beesHealer = require("../../choreographies/beesHealer");
 var fs = require("fs");
-
+var path = require("path");
 //TODO: prevent a class of race condition type of errors by signaling with files metadata to the watcher when it is safe to consume
 
 function FolderMQ(folder, callback){
-    var path = require("path");
     folder = path.normalize(folder);
 
 
@@ -20,7 +19,7 @@ function FolderMQ(folder, callback){
     });
 
     function mkFileName(swarmRaw){
-        return folder + "/" + swarmRaw.meta.swarmId + "."+swarmRaw.meta.swarmTypeName;
+        return path.normalize(folder + "/" + swarmRaw.meta.swarmId + "."+swarmRaw.meta.swarmTypeName);
     }
 
     this.getHandler = function(){
@@ -36,11 +35,18 @@ function FolderMQ(folder, callback){
                 if(!callback){
                     callback = $$.defaultErrorHandlingImplementation;
                 }
-                console.log("Scriu");
 
                 beesHealer.asJSON(swarm,null, null, function(err, res){
-                    console.log("Chiar scriu");
-                    fs.writeFile(mkFileName(res),J(res), callback);
+                    writeFile(mkFileName(res), J(res), callback);
+                });
+            },
+            sendSwarmForExecution: function(swarm, callback){
+                if(!callback){
+                    callback = $$.defaultErrorHandlingImplementation;
+                }
+
+                beesHealer.asJSON(swarm, swarm.meta.phaseName, null, function(err, res){
+                    writeFile(mkFileName(res), J(res), callback);
                 });
             }
         }
@@ -48,7 +54,7 @@ function FolderMQ(folder, callback){
 
     this.registerConsumer = function(callback){
         if(consumer){
-            throw new Error("Only one consumer is allowed!");
+            throw new Error("Only one consumer is allowed! "+folder);
         }
         consumer = callback;
         fs.mkdir(folder, function(err,res){
@@ -61,11 +67,18 @@ function FolderMQ(folder, callback){
     var producer = null;
 
     function consumeMessage(filename){
-        var fullPath = folder+"/"+filename;
+        var fullPath = path.normalize(folder+"/"+filename);
         fs.readFile(fullPath, "utf8", function(err, data){
             if(!err){
-                consumer(JSON.parse(data));
-                fs.unlink(fullPath, function(err, res){if(err) throw err});
+                if(data!==""){
+                    try{
+                        var message = JSON.parse(data);
+                    }catch(err){
+                        console.log(err, data);
+                    }
+                    consumer(message);
+                    fs.unlink(fullPath, function(err, res){if(err) throw err});
+                }
             }
         });
     }
@@ -75,24 +88,60 @@ function FolderMQ(folder, callback){
             if(err){
                 $$.errorHandler.error(err)
             } else {
-                files.forEach(consumeMessage);
                 watchFolder();
+                files.forEach(function(filename){
+                    if(filename.indexOf(in_progress)==-1){
+                        consumeMessage(filename);
+                    }
+                });
             }
-
         });
+    }
+
+    var in_progress = ".in_progress";
+    function writeFile(filename, content, callback){
+        var tmpFilename = filename+in_progress;
+        fs.writeFile(tmpFilename, content, function(error, res){
+            if(!error){
+                fs.rename(tmpFilename, filename, function(err, result){
+                    if(err){
+                        throw err;
+                    }else{
+                        callback(err, res);
+                    }
+                })
+            }else{
+                throw error;
+            }
+        });
+    }
+
+    var alreadyKnownChanges = {};
+
+    function alreadyFiredChanges(filename, change){
+        var res = false;
+        if(alreadyKnownChanges[filename]){
+           res = true;
+        }else{
+            alreadyKnownChanges[filename] = change;
+        }
+
+        return res;
     }
 
     function watchFolder(){
         fs.watch(folder, function(eventType, filename){
             //console.log("Watching:", eventType, filename);
-            if (filename && eventType == "change") {
-                consumeMessage(filename);
+
+            if (filename && (eventType == "change" || eventType == "rename")) {
+                if(filename.indexOf(in_progress)==-1 && !alreadyFiredChanges(filename, eventType)){
+                    consumeMessage(filename);
+                }
             }
         });
     }
 }
 
 exports.getFolderQueue = function(folder, callback){
-
     return new FolderMQ(folder, callback);
 };
